+++ src/main/java/binbean/binbean_BE/auth/filter/JwtVerificationFilter.java
@@ -2,10 +2,8 @@ package binbean.binbean_BE.auth.filter;
 
 import binbean.binbean_BE.auth.JwtTokenProvider;
 import binbean.binbean_BE.auth.UserDetailsImpl;
-import binbean.binbean_BE.constants.Constants;
 import binbean.binbean_BE.constants.Constants.LoggingMsg;
-import binbean.binbean_BE.constants.Constants.URL;
-import binbean.binbean_BE.service.auth.AuthService;
+import binbean.binbean_BE.service.AuthService;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.JwtException;
 import jakarta.servlet.FilterChain;
@@ -13,7 +11,6 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.Set;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
