+++ src/main/java/binbean/binbean_BE/config/SecurityConfig.java
@@ -5,12 +5,10 @@ import binbean.binbean_BE.auth.filter.JwtExceptionFilter;
 import binbean.binbean_BE.auth.filter.JwtUsernamePasswordAuthFilter;
 import binbean.binbean_BE.auth.filter.JwtVerificationFilter;
 import binbean.binbean_BE.auth.filter.UrlBasedAuthenticationFilter;
-import binbean.binbean_BE.constants.Constants;
 import binbean.binbean_BE.constants.Constants.URL;
 import binbean.binbean_BE.infra.RedisService;
-import binbean.binbean_BE.service.auth.AuthService;
+import binbean.binbean_BE.service.AuthService;
 import java.util.List;
-import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
