+++ src/main/java/binbean/binbean_BE/entity/Cafe.java
@@ -2,14 +2,13 @@ package binbean.binbean_BE.entity;
 
 import binbean.binbean_BE.dto.response.CafeInfoResponse;
 import binbean.binbean_BE.dto.response.ReviewResponse;
-import binbean.binbean_BE.entity.user.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import java.util.List;
 import lombok.Builder;
@@ -29,10 +28,6 @@ public class Cafe {
     @ManyToOne
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
-
-    @OneToOne
-    @JoinColumn(name = "business_hours_id", nullable = false)
-    private BusinessHours businessHours;
 
     @Column(name = "cafe_name", nullable = false)
     private String cafeName;
@@ -67,11 +62,11 @@ public class Cafe {
     protected Cafe() {}
 
     @Builder
-    public Cafe(Long id, BusinessHours businessHours, String cafeName, String cafeAddress,
+    public Cafe(Long id, User user, String cafeName, String cafeAddress,
         double latitude, double longitude, String cafePhone, String cafeDescription, int wifiAvailable, int petAvailable,
         int kidsAvailable, int chargeAvailable) {
         this.id = id;
-        this.businessHours = businessHours;
+        this.user = user;
         this.cafeName = cafeName;
         this.cafeAddress = cafeAddress;
         this.latitude = latitude;
