+++ src/main/java/binbean/binbean_BE/exception/GlobalExceptionHandler.java
@@ -1,5 +1,6 @@
 package binbean.binbean_BE.exception;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
