+++ src/main/java/binbean/binbean_BE/service/CafeService.java
@@ -9,14 +9,11 @@ import binbean.binbean_BE.dto.response.ReviewResponse;
 import binbean.binbean_BE.entity.BusinessHours;
 import binbean.binbean_BE.entity.Cafe;
 import binbean.binbean_BE.entity.CafeImg;
-import binbean.binbean_BE.entity.Review;
 import binbean.binbean_BE.entity.floor_plan.FloorPlan;
 import binbean.binbean_BE.exception.NotFoundException;
 import binbean.binbean_BE.repository.BusinessHoursRepository;
 import binbean.binbean_BE.repository.CafeImgRepository;
 import binbean.binbean_BE.repository.CafeRepository;
-import binbean.binbean_BE.repository.ReviewImgRepository;
-import binbean.binbean_BE.repository.ReviewRepository;
 import binbean.binbean_BE.repository.floor_plan.BorderLineRepository;
 import binbean.binbean_BE.repository.floor_plan.CounterRepository;
 import binbean.binbean_BE.repository.floor_plan.DoorRepository;
